laptop/requirements.txt
# Project-Cortex v2.0 - Laptop Server Dependencies
# Install with: pip install -r laptop/requirements.txt

# ============ CORE DEPENDENCIES (REQUIRED) ============

# GUI Framework
PyQt6>=6.6.0

# WebSocket Server
websockets>=12.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# ============ OPTIONAL DEPENDENCIES (FastAPI Server) ============

# Install with: pip install -r laptop/requirements.txt[api]

# REST API Framework
fastapi>=0.108.0
uvicorn[standard]>=0.25.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# ============ DEVELOPMENT DEPENDENCIES (Optional) ============

# Code Quality
# pylint>=3.0.0
# black>=23.0.0
# mypy>=1.7.0

# Testing
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# pytest-qt>=4.2.0
